PPC swizzle
