+++ include/eve/detail/function/patterns/matcher.hpp
@@ -56,20 +56,12 @@ namespace eve::detail
       {
         return r.slice(upper_)[ slide_pattern<cd/2,sz>(p) ];
       }
-      // We're swizzling the a full aggregate
-      else if constexpr( has_aggregated_abi_v<Wide> )
-      {
-        puts("FUCKED UP");
-        return that_t{};
-      }
       else
       {
         // We select the proper shuffler/swizzler
         using found = typename type_at< find( as_pattern<sz>(Pattern{}), as_<Wide>(), as_<that_t>())
                                       , types<Matcher...>
                                       >::type;
-
-        std::cout << typeid(found).name() << "\n";
 
         // We call the associated processing function
         return do_swizzle ( delay_t{}, EVE_CURRENT_API{}
