+++ include/eve/detail/function/patterns/slide_left.hpp
@@ -25,47 +25,46 @@ namespace eve::detail
   //================================================================================================
   struct slide_left
   {
-    template<typename In, typename Out, std::ptrdiff_t... I>
-    static constexpr auto check(pattern_<I...>, as_<In>, as_<Out>)  noexcept
+    template<typename In, typename Out, std::ptrdiff_t I0, std::ptrdiff_t... I>
+    static constexpr auto check(pattern_<I0,I...>, as_<In>, as_<Out>)  noexcept
     {
-      constexpr std::ptrdiff_t c  = cardinal_v<In>;
-      constexpr std::ptrdiff_t sz = cardinal_v<Out>;
-      constexpr pattern_<I...> p;
+      std::ptrdiff_t c  = cardinal_v<In>;
+      std::ptrdiff_t sz = cardinal_v<Out>;
+      pattern_<I0,I...> p;
 
-      if(c==1) return false;
+      // No slide of 1
+      if(c==1 || sz ==1) return false;
+
+      // We must starts with at least 1
+      auto p0 = p(0,sz);
+      if(p0 < 1) return false;
 
       // Valid pattern
       // [ k (k+1<N ? k+1 : -1) .. (k+sz-1<N ? k+sz-1 : -1) ]  1 <= k <= N-1
-      auto p0 = p(0,c);
-
-      // We must starts with at least 1
-      if(p0 < 1) return false;
-
       int  i  = 1;
-      while(i<sz && p(i,c) != -1)
+      while(i<sz && p(i,sz) != -1)
       {
-        auto q = p(i,c);
-        if(q != p0+i)  return false;
+        if(p(i,sz) != p0+i)  return false;
         i++;
       }
-      if(i != sz)
-      {
-        if(c ==sz)
-        {
-          while( i<sz )
-          {
-            auto q = p(i,c);
-            if(q!=-1) return false;
-            i++;
-          }
-        }
-        else
-        {
-          return false;
-        }
-      }
 
-      return true;
+      if( i<sz && p(i,sz) == -1 )
+      {
+        // -1 means we ran out of original cardinal indexing
+        if(p(i-1,sz) != c-1) return false;
+
+        while( i<sz )
+        {
+          if(p(i,sz) != -1) return false;
+          i++;
+        }
+
+        return true;
+      }
+      else
+      {
+        return true;
+      }
     }
   };
 }
