+++ include/eve/detail/function/swizzle.hpp
@@ -17,9 +17,9 @@
 #  include <eve/detail/function/simd/x86/swizzle.hpp>
 #endif
 
-// #if defined(EVE_HW_POWERPC)
-// #  include <eve/detail/function/simd/ppc/swizzle.hpp>
-// #endif
+#if defined(EVE_HW_POWERPC)
+#  include <eve/detail/function/simd/ppc/swizzle.hpp>
+#endif
 
 #if defined(EVE_HW_ARM)
 #  include <eve/detail/function/simd/arm/neon/swizzle.hpp>
